backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
google-generativeai==0.3.2
PyPDF2==3.0.1
python-docx==1.1.0
Pillow==10.1.0
python-decouple==3.8
django-filter==23.5

# ML Libraries for Policy Comparison
scikit-learn==1.7.1
numpy==2.2.1

# Document Processing Libraries
PyPDF2==3.0.1
python-docx==1.1.0
